+++ ReleaseNotes.md
@@ -5,10 +5,12 @@
 ### AutoSpectre
 
 * TextPrompt has three new properties for defining choices.
+* DefaultValueSource added on TextPrompt
 
 ### AutoSpectre.SourceGeneration
 
 * Based on the new properties in TextPrompt, the source generation has been updated to support these new properties.
+* New solution for defining DefaultValueSource. Old way discarded. 
 
 ## 0.6.0
 
@@ -23,3 +25,5 @@
 * Fix to allow inheritance for form classes (classes decorated with AutoSpectreForm)
 * Allow setting culture to use for source generation. The culture is initalized as a culture variable. Per default it will be the CurrentUICulture
 * Added ability to point to a method to initialize anohter referenced form. This is useful if that class has a constructor that is not empty.
+
+